Move activation event handling into `EpropArchivingNode`
