+++ models/eprop_iaf_adapt.cpp
@@ -86,7 +86,6 @@ eprop_iaf_adapt::Parameters_::Parameters_()
   , kappa_( 0.97 )
   , kappa_reg_( 0.97 )
   , eprop_isi_trace_cutoff_( 1000.0 )
-  , activation_interval_( 3000.0 )
 {
 }
 
@@ -137,7 +136,6 @@ eprop_iaf_adapt::Parameters_::get( Dictionary& d ) const
   d[ names::kappa ] = kappa_;
   d[ names::kappa_reg ] = kappa_reg_;
   d[ names::eprop_isi_trace_cutoff ] = eprop_isi_trace_cutoff_;
-  d[ names::activation_interval ] = activation_interval_;
 }
 
 double
@@ -177,7 +175,6 @@ eprop_iaf_adapt::Parameters_::set( const Dictionary& d, Node* node )
   update_value_param( d, names::kappa, kappa_, node );
   update_value_param( d, names::kappa_reg, kappa_reg_, node );
   update_value_param( d, names::eprop_isi_trace_cutoff, eprop_isi_trace_cutoff_, node );
-  update_value_param( d, names::activation_interval, activation_interval_, node );
 
   if ( adapt_beta_ < 0 )
   {
@@ -227,17 +224,6 @@ eprop_iaf_adapt::Parameters_::set( const Dictionary& d, Node* node )
   if ( kappa_reg_ < 0.0 or kappa_reg_ > 1.0 )
   {
     throw BadProperty( "Firing rate low-pass filter for regularization kappa_reg from range [0, 1] required." );
-  }
-
-  if ( activation_interval_ < 0 )
-  {
-    throw BadProperty( "Interval between activations activation_interval ≥ 0 required." );
-  }
-
-  if ( eprop_isi_trace_cutoff_ < 0.0 or eprop_isi_trace_cutoff_ > activation_interval_ )
-  {
-    throw BadProperty(
-      "Computation cutoff of eprop trace 0 ≤ eprop trace eprop_isi_trace_cutoff ≤ activation_interval required." );
   }
   return delta_EL;
 }
@@ -310,7 +296,6 @@ eprop_iaf_adapt::pre_run_hook()
 
   V_.RefractoryCounts_ = Time( Time::ms( P_.t_ref_ ) ).get_steps();
   V_.eprop_isi_trace_cutoff_steps_ = Time( Time::ms( P_.eprop_isi_trace_cutoff_ ) ).get_steps();
-  V_.activation_interval_steps_ = Time( Time::ms( P_.activation_interval_ ) ).get_steps();
 
   // calculate the entries of the propagator matrix for the evolution of the state vector
 
@@ -361,7 +346,7 @@ eprop_iaf_adapt::update( Time const& origin, const long from, const long to )
       S_.r_ = V_.RefractoryCounts_;
       set_last_event_time( t );
     }
-    else if ( get_last_event_time() > 0 and t - get_last_event_time() >= V_.activation_interval_steps_ )
+    else if ( get_last_event_time() > 0 and t - get_last_event_time() >= activation_interval_steps_ )
     {
       SpikeEvent se;
       se.set_activation();
