+++ models/eprop_iaf_adapt_bsshslm_2020.h
@@ -388,9 +388,6 @@ private:
     //! Spike threshold voltage relative to the leak membrane potential (mV).
     double V_th_;
 
-    //! Interval between two activations.
-    long activation_interval_;
-
     //! Default constructor.
     Parameters_();
 
@@ -478,9 +475,6 @@ private:
 
     //! Total refractory steps.
     int RefractoryCounts_;
-
-    //! Time steps of activation interval.
-    long activation_interval_steps_;
   };
 
   //! Get the current value of the membrane voltage.
@@ -603,6 +597,7 @@ eprop_iaf_adapt_bsshslm_2020::handles_test_event( DataLoggingRequest& dlr, size_
 inline void
 eprop_iaf_adapt_bsshslm_2020::get_status( Dictionary& d ) const
 {
+  EpropArchivingNode::get_status( d );
   P_.get( d );
   S_.get( d, P_ );
   d[ names::recordables ] = recordablesMap_.get_list();
@@ -611,6 +606,7 @@ eprop_iaf_adapt_bsshslm_2020::get_status( Dictionary& d ) const
 inline void
 eprop_iaf_adapt_bsshslm_2020::set_status( const Dictionary& d )
 {
+  EpropArchivingNode::set_status( d );
   // temporary copies in case of errors
   Parameters_ ptmp = P_;
   State_ stmp = S_;
