+++ nestkernel/archiving_node.h
@@ -96,12 +96,6 @@ public:
   void set_status( const Dictionary& d ) override;
 
 protected:
-  //! Interval between two activations in steps.
-  long activation_interval_;
-
-  //! Time of last spike or activation event in steps.
-  long last_event_time_;
-
   /**
    * Record spike history
    */
